server/requirements.txt
attrs==22.2.0
bcrypt==4.0.1
click==8.1.3
flasgger==0.9.5
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
greenlet==2.0.1
importlib-metadata==5.2.0
importlib-resources==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
mistune==2.0.4
pkgutil_resolve_name==1.3.10
PyJWT==2.6.0
pyrsistent==0.19.2
python-decouple==3.7
python-dotenv==0.21.0
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
zipp==3.11.0
